Fixes #28: extracted an interface from the aggregate root repository.
